control work 1_2

+++ src/Employee.java
@@ -39,8 +39,8 @@ public class Employee {
         this.department = department;
     }
 
-    public void setSalary(int salary) {
-        this.salary = salary;
+    public void setSalary(double salary) {
+        this.salary = (int) salary;
     }
 
     @Override
